refactor(geo): Move class into better section to encapsulate their behaviour Closes #2046
