+++ packages/geoview-core/src/api/event-processors/event-processor-children/legend-event-processor.ts
@@ -7,7 +7,7 @@ import {
   isVectorLegend,
   isWmsLegend,
 } from '@/geo/layer/geoview-layers/abstract-geoview-layers';
-import { TypeLegendResultSetEntry } from '@/geo/utils/legends-layer-set';
+import { TypeLegendResultSetEntry } from '@/geo/layer/layer-sets/legends-layer-set';
 import { ILayerState } from '@/core/stores/store-interface-and-intial-values/layer-state';
 import { getLocalizedValue } from '@/core/utils/utilities';
 import { AbstractEventProcessor } from '@/api/event-processors/abstract-event-processor';
