+++ packages/geoview-core/src/core/components/data-table/data-panel.tsx
@@ -17,7 +17,7 @@ import { useAppGuide } from '@/core/stores/store-interface-and-intial-values/app
 import { LayerListEntry, Layout } from '@/core/components/common';
 import { logger } from '@/core/utils/logger';
 import { useFeatureFieldInfos } from './hooks';
-import { TypeFieldEntry, TypeLayerData } from '@/geo/utils/layer-set';
+import { TypeFieldEntry, TypeLayerData } from '@/geo/layer/layer-sets/layer-set';
 import { getSxClasses } from './data-table-style';
 import { LAYER_STATUS, TABS } from '@/core/utils/constant';
