+++ packages/geoview-core/src/core/components/details/feature-info-new.tsx
@@ -8,7 +8,7 @@ import { useDetailsCheckedFeatures, useDetailsStoreActions } from '@/core/stores
 import { useMapStoreActions } from '@/core/stores/store-interface-and-intial-values/map-state';
 import { logger } from '@/core/utils/logger';
 import { delay } from '@/core/utils/utilities';
-import { TypeFeatureInfoEntry, TypeFieldEntry, TypeGeometry } from '@/geo/utils/layer-set';
+import { TypeFeatureInfoEntry, TypeFieldEntry, TypeGeometry } from '@/geo/layer/layer-sets/layer-set';
 
 import { FeatureInfoTable } from './feature-info-table';
 import { getSxClasses } from './details-style';
