+++ packages/geoview-core/src/geo/index.ts
@@ -24,12 +24,17 @@ export * from './layer/geoview-layers/esri-layer-common';
 export * from './layer/other/geocore';
 export * from './layer/geometry/geometry-types';
 export * from './layer/geometry/geometry';
+export * from './layer/layer-sets/all-feature-info-layer-set';
+export * from './layer/layer-sets/feature-info-layer-set';
+export * from './layer/layer-sets/hover-feature-info-layer-set';
+export * from './layer/layer-sets/layer-set';
+export * from './layer/layer-sets/legends-layer-set';
 export * from './layer/layer';
 export * from './map/map-schema-types';
 export * from './map/map-viewer';
-export * from './projection/projection';
-export * from './renderer/esri-renderer';
-export * from './renderer/geoview-renderer-types';
-export * from './renderer/geoview-renderer';
-export * from './utils/feature-highlight';
+export * from './map/feature-highlight';
+export * from './utils/projection';
+export * from './utils/renderer/esri-renderer';
+export * from './utils/renderer/geoview-renderer-types';
+export * from './utils/renderer/geoview-renderer';
 export * from './utils/utilities';
