+++ packages/geoview-core/src/geo/layer/geoview-layers/vector/abstract-geoview-vector.ts
@@ -23,16 +23,16 @@ import {
 import { api } from '@/app';
 import { getLocalizedValue } from '@/core/utils/utilities';
 import { getMinOrMaxExtents } from '@/geo/utils/utilities';
-import { NodeType } from '@/geo/renderer/geoview-renderer-types';
+import { NodeType } from '@/geo/utils/renderer/geoview-renderer-types';
 import { MapEventProcessor } from '@/api/event-processors/event-processor-children/map-event-processor';
 import { logger } from '@/core/utils/logger';
 import { CSV } from './csv';
 import { VectorLayerEntryConfig } from '@/core/utils/config/validation-classes/vector-layer-entry-config';
 import { AbstractBaseLayerEntryConfig } from '@/core/utils/config/validation-classes/abstract-base-layer-entry-config';
-import { TypeFeatureInfoEntry } from '@/geo/utils/layer-set';
+import { TypeFeatureInfoEntry } from '@/geo/layer/layer-sets/layer-set';
 import { Cast } from '@/core/types/global-types';
 import { AppEventProcessor } from '@/api/event-processors/event-processor-children/app-event-processor';
-import { analyzeLayerFilter, getFeatureStyle } from '@/geo/renderer/geoview-renderer';
+import { analyzeLayerFilter, getFeatureStyle } from '@/geo/utils/renderer/geoview-renderer';
 
 /* *******************************************************************************************************************************
  * AbstractGeoViewVector types
