+++ packages/geoview-core/src/geo/utils/renderer/esri-renderer.ts
